Enabled background running and tray storage
Converting the python file to .pyw causes powershell windows to appear, as an artifact of tinyWinToast. Converting to an exe causes an error to be thrown, and this is because tinyWinToast creates a .ps1 file, and for whatever reason, it cannot create that file when it's converted.

I may try and ditch tinyWinToast because of this issue so others can use it, but for now, things will remain the same.