Added local paths
I wanted to make this open at startup for my Linux machine, but doing so caused the files to be added to my /home/ folder, so I made it so the paths were entirely localized to wherever you keep the python file. I don't know if this works on Windows yet as I haven't tried it, but I coded it in such a way where hopefully it should work without issue.