Added Submission Detail
It now lets you know if you've submitted something for the assignment or not. This is still in the early stages and I may also add a config option to skip showing assignments if they have submissions or something, depends on how I feel about it, and - if this for whatever reason gains traction - what people want.