Removed Submission Check Functionality
Unfortunately, this check was false; upon using this myself and seeing an incorrect state of submission, I checked Canvas' API and found out that "has_submitted_submissions" does not mean "has the user submitted anything" and rather means "has *ANY* student submitted anything," which unfortunately is rather broad and undermines the point of this. I can't find any consistent way to check for submissions as of right now.