Shifted the time further "back"
I have yet to find an effective way of changing the time received from the Canvas API to the correct time zone, so I'm currently shifting the reported time for assignments' due dates back by 8 hours. I'll review this later to see if there might be some way to correctly alter the notifications and time for each time zone.