Added URL button to Windows
Officially added the URL button on Windows (Linux will likely never receive this :/) - it links you directly to the assignment that is being notified, that way you don't have to take notes on every assignment being notified and instead can just go straight to it.

I had to perform a workaround since the library that sends notifications is pretty inconsistent at times, so see the code for further details on that.